+++ parser/config.go
@@ -146,6 +146,7 @@ var SupportedFunctions = map[string]int{
 	"omega":     1,
 	"tau":       1,
 	"mobius":    1,
+	"mertens":   1,
 	"bigomega":  1,
 	"sumofprimes": 1,
 	"lambertw":  1,